+++ tests/Base/ResponseBuilderTestCaseBase.php
@@ -20,7 +20,7 @@ use Symfony\Component\HttpFoundation\Response as HttpResponse;
 /**
  * Class ResponseBuilderTestCase
  */
-class ResponseBuilderTestCaseBase extends \Orchestra\Testbench\TestCase
+abstract class ResponseBuilderTestCaseBase extends TestCaseBase
 {
 	/**
 	 * @return ErrorCode
@@ -272,98 +272,5 @@ class ResponseBuilderTestCaseBase extends \Orchestra\Testbench\TestCase
 		                                             $data, $http_code, $lang_args, $headers]);
 	}
 
-
-
-	// ----------------------------------------------------------------------------
-
-
-//	/**
-//	 * Returns instance of your API codes class. Sufficient implementation of this method
-//	 * for most of the cases is just:
-//	 *
-//	 *   return new \App\ApiCodes();
-//	 *
-//	 * where \App\ApiCodes matches your codes class
-//	 *
-//	 * @return \MarcinOrlowski\ResponseBuilder\ErrorCode
-//	 */
-//	abstract public function getApiCodesObject();
-//
-//	/**
-//	 * return object of your API codes class usually just:
-//	 *
-//	 *   return '\App\ApiCodes';
-//	 *
-//	 * or (PHP5.5+ only!)
-//	 *
-//	 *   return \App\ApiCodes::class;
-//	 *
-//	 * NOTE: MUST start with the "\"!
-//	 *
-//	 * @return string
-//	 */
-//	abstract public function getApiCodesClassName();
-
-	/**
-	 * Returns ErrorCode constant name referenced by its value
-	 *
-	 * @param $error_code
-	 *
-	 * @return int|null|string
-	 */
-	protected function resolveConstantFromCode($error_code) {
-		/** @var \MarcinOrlowski\ResponseBuilder\ErrorCode $api_codes_class_name */
-		$api_codes_class_name = $this->getApiCodesClassName();
-		/** @var array $const */
-		$const = $api_codes_class_name::getErrorCodeConstants();
-		$name = null;
-		foreach( $const as $const_name => $const_value ) {
-			if( is_int($const_value) && ($const_value === $error_code) ) {
-				$name = $const_name;
-				break;
-			}
-		}
-
-		return ($name === null) ? "??? ({$error_code})" : $name;
-	}
-
-	/**
-	 * Helper to let test protected/private methods
-	 *
-	 * Usage example:
-	 * ----------------
-	 *   $method = $this->getProtectedMethod('App\Foo', 'someMethod');
-	 *   $obj = new \App\Foo();
-	 *   $result = $method->invokeArgs($obj, ...);
-	 *
-	 * @param string $class_name  name of the class method belongs to, i.e. "Bar". Can be namespaced i.e. "Foo\Bar" (no starting backslash)
-	 * @param string $method_name method name to call
-	 *
-	 * @return \ReflectionMethod
-	 */
-	protected function getProtectedMethod($class_name, $method_name)
-	{
-		$class = new \ReflectionClass($class_name);
-		$method = $class->getMethod($method_name);
-		$method->setAccessible(true);
-
-		return $method;
-	}
-
-	/**
-	 * Generates random string, with optional prefix
-	 *
-	 * @param string $prefix
-	 *
-	 * @return string
-	 */
-	protected function getRandomString($prefix = null)
-	{
-		if ($prefix !== null) {
-			$prefix = "{$prefix}_";
-		}
-
-		return $prefix . md5(uniqid(mt_rand(), true));
-	}
 
 }
